exit

+++ lib/libft/ft_atul.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atul.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kdaniely <kdaniely@student.42.fr>          +#+  +:+       +#+        */
+/*   By: dohanyan <dohanyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/07 18:13:56 by dohanyan          #+#    #+#             */
-/*   Updated: 2023/06/12 14:27:23 by kdaniely         ###   ########.fr       */
+/*   Updated: 2023/06/17 20:14:22 by dohanyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdio.h>
+
 #define OFFSET '0'
 
 static short	is_num(const char *chr)
@@ -39,9 +41,9 @@ static short	is_whitespace(const char *chr)
 */
 unsigned long	ft_atul(const char *str)
 {
-	long long	num;
-	short		sign;
-	const char	*c;
+	long long		num;
+	unsigned long	sign;
+	const char		*c;
 
 	sign = 1;
 	c = str;
@@ -62,5 +64,5 @@ unsigned long	ft_atul(const char *str)
 		num += (int)(*c - OFFSET);
 		c++;
 	}
-	return ((unsigned long)(sign * num));
+	return (sign * num);
 }
